Fixed issues in valve's code

+++ game/client/detailobjectsystem.cpp
@@ -1515,7 +1515,7 @@ void CDetailObjectSystem::LevelInitPostEntity()
 		IMaterial *pMat = m_DetailSpriteMaterial;
 
 		// adjust for non-square textures (cropped)
-		float flRatio = pMat->GetMappingWidth() / pMat->GetMappingHeight();
+		float flRatio = pMat->GetMappingWidth() / (float)pMat->GetMappingHeight();
 		if ( flRatio > 1.0 )
 		{
 			for( int i = 0; i < m_DetailSpriteDict.Count(); i++ )
